Move createMoteType() to ExtensionManager
Mote types are registered like plugins,
so this functionality belongs in this class.
