+++ java/org/contikios/cooja/ExtensionManager.java
@@ -30,7 +30,13 @@
 
 package org.contikios.cooja;
 
+import java.lang.reflect.InvocationTargetException;
 import java.util.LinkedHashMap;
+import org.contikios.cooja.contikimote.ContikiMoteType;
+import org.contikios.cooja.motes.DisturberMoteType;
+import org.contikios.cooja.motes.ImportAppMoteType;
+import org.contikios.cooja.mspmote.SkyMoteType;
+import org.contikios.cooja.mspmote.Z1MoteType;
 import org.contikios.cooja.mspmote.plugins.MspCLI;
 import org.contikios.cooja.mspmote.plugins.MspCodeWatcher;
 import org.contikios.cooja.mspmote.plugins.MspCycleWatcher;
@@ -98,5 +104,29 @@ public class ExtensionManager {
       }
     }
     return clazz;
+  }
+
+  /** Create a mote of a certain class, returns null on failure. */
+  public static MoteType createMoteType(Cooja cooja, String name) {
+    switch (name) {
+      case "org.contikios.cooja.motes.ImportAppMoteType": return new ImportAppMoteType();
+      case "org.contikios.cooja.motes.DisturberMoteType": return new DisturberMoteType();
+      case "org.contikios.cooja.contikimote.ContikiMoteType": return new ContikiMoteType(cooja);
+      case "org.contikios.cooja.mspmote.SkyMoteType": return new SkyMoteType();
+      case "org.contikios.cooja.mspmote.Z1MoteType": return new Z1MoteType();
+    }
+    Class<? extends MoteType> moteType = null;
+    for (var clazz : cooja.getRegisteredMoteTypes()) {
+      if (name.equals(clazz.getName())) {
+        moteType = clazz;
+        break;
+      }
+    }
+    if (moteType == null) return null;
+    try {
+      return moteType.getConstructor().newInstance();
+    } catch (InstantiationException | IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
+      return null;
+    }
   }
 }
